requirements.txt
streamlit
markdown
python-docx
beautifulsoup4
pypandoc
Pillow
